+++ Lab1/Assets/Scenes/MainMenu.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 49bcdcd2226fb6249bd73b1954748350
+guid: 08e4753f7397ccf4dae92e3645c582be
 DefaultImporter:
   externalObjects: {}
   userData:
