+++ Lab1/Assets/Scripts/Controller.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: edcdf720a2ad4a44b84930c5278f1661
+guid: 4f6c34acfbb74744bbb03406ae9cbdb9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
