+++ Lab1/Assets/Scripts/EnemyCollision.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 808dee837bb30d944bdfa0ca40535f44
+guid: e9a4f7830a0323a46a134a5dad71a9bd
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
