+++ Lab1/Assets/Scripts/EnemyController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7a9e0635bb098fc49afb843403138080
+guid: 6471450279c228e4ba85fa99f8ee6e89
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
