+++ Lab1/Assets/Scripts/GameState.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 015733df8d2b9ab49863729addf0bf42
+guid: acdec7e41de0e63409f333984b3fefeb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
