--- Lab1/Assets/Scripts/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 682c15b0e509ba446a4512a9c671d480
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
