--- Lab1/Assets/Scripts/Player/PlayerAttack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 800b74f00404122488ba9f66d6216bf1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
