+++ Lab1/Assets/Scripts/Title.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 905ebc109e544b246bc61acb9a197c27
+guid: 2cd350dc2552ede4da6f2558f63fe55b
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
